Move mongo import to start of imports
