+++ udata/api/oauth2.py
@@ -32,10 +32,10 @@ from flask_security.utils import verify_password
 from werkzeug.exceptions import Unauthorized
 from werkzeug.security import gen_salt
 
+from udata.mongo import db
 from udata.app import csrf
 from udata.auth import current_user, login_required, login_user
 from udata.i18n import I18nBlueprint, lazy_gettext as _
-from udata.mongo import db
 from udata.core.user.models import User
 from udata.core.storages import images, default_image_basename
