+++ udata/core/activity/api.py
@@ -2,9 +2,9 @@ import logging
 
 from mongoengine.errors import DoesNotExist
 
+from udata.mongo import db
 from udata.api import api, API, fields
 from udata.core.activity.models import Activity
-from udata.mongo import db
 
 from udata.core.user.api_fields import user_ref_fields
 from udata.core.organization.api_fields import org_ref_fields
