+++ udata/core/badges/tests/test_model.py
@@ -1,5 +1,5 @@
-from udata.auth import login_user
 from udata.mongo import db
+from udata.auth import login_user
 from udata.tests import TestCase, DBTestMixin
 from udata.core.user.factories import UserFactory
