+++ udata/core/dataset/models.py
@@ -16,11 +16,11 @@ from werkzeug.utils import cached_property
 import requests
 from typing import Optional, Tuple
 
+from udata.mongo import db
 from udata.app import cache
 from udata.core import storages
 from udata.frontend.markdown import mdstrip
 from udata.models import WithMetrics, BadgeMixin, SpatialCoverage, FieldValidationError
-from udata.mongo import db
 from udata.i18n import lazy_gettext as _
 from udata.utils import get_by, hash_url, to_naive_datetime
 from udata.uris import ValidationError, endpoint_for
