+++ udata/core/organization/activities.py
@@ -1,8 +1,8 @@
 from flask_security import current_user
 
 from udata.i18n import lazy_gettext as _
-from udata.models import Organization, Activity
 from udata.mongo import db
+from udata.models import Organization, Activity
 
 
 __all__ = (
