+++ udata/core/organization/metrics.py
@@ -1,5 +1,5 @@
-from udata.models import Dataset, Reuse, Organization
 from udata.mongo import db
+from udata.models import Dataset, Reuse, Organization
 
 
 @Dataset.on_create.connect
