+++ udata/core/spam/tests/test_spam.py
@@ -1,8 +1,8 @@
 import logging
 import pytest
-from udata.tests import TestCase
-from udata.tests.helpers import assert_not_emit
+
 from udata.mongo import db
+from udata.tests import TestCase
 
 from ..models import POTENTIAL_SPAM, SpamMixin
