+++ udata/core/spatial/models.py
@@ -2,10 +2,10 @@ from flask import current_app
 from werkzeug.local import LocalProxy
 from werkzeug.utils import cached_property
 
+from udata.mongo import db
 from udata.app import cache
 from udata.uris import endpoint_for
 from udata.i18n import _, L_, get_locale, language
-from udata.mongo import db
 
 from . import geoids
