+++ udata/core/spatial/tests/test_fields.py
@@ -4,8 +4,8 @@ from datetime import timedelta
 
 from werkzeug.datastructures import MultiDict
 
-from udata.forms import Form
 from udata.mongo import db
+from udata.forms import Form
 from udata.tests import TestCase
 from udata.tests.helpers import assert_json_equal
 from udata.utils import faker
