+++ udata/core/topic/models.py
@@ -1,7 +1,8 @@
 from flask import url_for
 from mongoengine.signals import pre_save
-from udata.models import SpatialCoverage
+
 from udata.mongo import db
+from udata.models import SpatialCoverage
 from udata.search import reindex
 from udata.tasks import as_task_param
