+++ udata/core/user/models.py
@@ -12,12 +12,12 @@ from mongoengine.signals import pre_save, post_save
 
 from werkzeug.utils import cached_property
 
+from udata.mongo import db
 from udata import mail
 from udata.uris import endpoint_for
 from udata.frontend.markdown import mdstrip
 from udata.i18n import lazy_gettext as _
 from udata.models import WithMetrics, Follow
-from udata.mongo import db
 from udata.core.discussions.models import Discussion
 from udata.core.storages import avatars, default_image_basename
