+++ udata/features/transfer/api.py
@@ -1,12 +1,12 @@
 from flask import request
 
+from udata.mongo import db
 from udata.api import api, fields, API, base_reference
 from udata.core.dataset.api_fields import dataset_ref_fields
 from udata.core.organization.api_fields import org_ref_fields
 from udata.core.reuse.api_fields import reuse_ref_fields
 from udata.core.user.api_fields import user_ref_fields
 from udata.models import User, Organization, Dataset, Reuse
-from udata.mongo import db
 from udata.utils import id_or_404
 
 from .actions import request_transfer, accept_transfer, refuse_transfer
