+++ udata/tests/forms/test_basic_fields.py
@@ -1,7 +1,7 @@
 from werkzeug.datastructures import MultiDict
 
-from udata.forms import Form, fields
 from udata.mongo import db
+from udata.forms import Form, fields
 from udata.tests import TestCase
 from udata.utils import faker
