+++ udata/tests/forms/test_current_user_field.py
@@ -3,13 +3,13 @@ from bson import ObjectId
 
 from werkzeug.datastructures import MultiDict
 
+from udata.mongo import db
 from udata.auth import login_user
 from udata.auth.forms import ExtendedLoginForm, ExtendedRegisterForm
 from udata.core.user.factories import UserFactory, AdminFactory
 from udata.forms import ModelForm, fields
 from udata.i18n import gettext as _
 from udata.models import User
-from udata.mongo import db
 from udata.tests import TestCase, DBTestMixin
 from udata.tests.helpers import security_gettext
