+++ udata/tests/forms/test_extras_fields.py
@@ -5,8 +5,8 @@ from uuid import UUID
 
 from werkzeug.datastructures import MultiDict
 
-from udata.forms import fields, ModelForm
 from udata.mongo import db
+from udata.forms import fields, ModelForm
 
 pytestmark = [
     pytest.mark.usefixtures('app')
