+++ udata/tests/forms/test_model_field.py
@@ -2,9 +2,9 @@ import pytest
 
 from werkzeug.datastructures import MultiDict
 
+from udata.mongo import db
 from udata.forms import ModelForm, fields
 from udata.i18n import gettext as _
-from udata.mongo import db
 
 
 pytestmark = [
