+++ udata/tests/forms/test_model_list_field.py
@@ -1,7 +1,7 @@
 from werkzeug.datastructures import MultiDict
 
-from udata.forms import ModelForm, fields
 from udata.mongo import db
+from udata.forms import ModelForm, fields
 from udata.tests import TestCase
 from udata.utils import faker
