+++ udata/tests/plugin.py
@@ -9,10 +9,10 @@ from flask.testing import FlaskClient
 from lxml import etree
 from werkzeug.urls import url_encode
 
+from udata.mongo import db
 from udata import settings
 from udata.app import create_app
 from udata.core.user.factories import UserFactory
-from udata.mongo import db
 
 from .helpers import assert200, assert_command_ok
