+++ udata/tests/test_owned.py
@@ -1,12 +1,10 @@
 from mongoengine import post_save
 
+from udata.mongo import db
 from udata.core.organization.factories import OrganizationFactory
 from udata.core.organization.models import Organization
 from udata.core.user.factories import UserFactory
 from udata.core.user.models import User
-
-from udata.mongo import db
-
 from udata.tests import TestCase, DBTestMixin
