+++ udata/tests/test_routing.py
@@ -5,11 +5,11 @@ from uuid import uuid4
 
 from flask import url_for
 
+from udata.mongo import db
 from udata import routing
 from udata.core.spatial.models import GeoZone
 from udata.core.spatial.factories import GeoZoneFactory
-from udata.mongo import db
-from udata.models.slug_fields import SlugFollow
+from udata.mongo.slug_fields import SlugFollow
 from udata.tests.helpers import assert200, assert404, assert_redirects
